Reorganize ranchers directory for AWS secrets manager
Reorganizes the ranchers directory to clear the way for AWS secrets manager to be used in most test cases.

Update eula.go

use airgap tunnel

remove tunnel changes
