--- tests/infrastructure/ranchers/defaults/defaults.yaml
@@ -1,79 +1,0 @@
-#######################
-# RANCHER CONFIG
-#######################
-rancher:
-  host: "<required>"
-  adminPassword: "<HB_RANCHER_ADMIN_PASSWORD>"
-  insecure: true
-  cleanup: false
-
-#######################
-# TERRAFORM CONFIG
-#######################
-terraform:
-  cni: "calico"
-  defaultClusterRoleForProjectMembers: "true"
-  downstreamClusterProvider: "<HB_PROVIDER_AMAZON>"
-  enableNetworkPolicy: false
-  provider: "<HB_PROVIDER_AMAZON>"
-  privateKeyPath: "<HB_SSH_PRIVATE_KEY_PATH>"
-  resourcePrefix: "auto-tfp"
-  privateRegistries:
-    url: "<HB_PRIVATE_REGISTRY_URL>"
-    username: "<HB_PRIVATE_REGISTRY_USERNAME>"
-    password: "<HB_PRIVATE_REGISTRY_PASSWORD>"
-    insecure: true
-    authConfigSecretName: "<HB_AUTH_CONFIG_SECRET_NAME>"
-    mirrorHostname: "<HB_PRIVATE_REGISTRY_MIRROR_HOSTNAME>"
-    mirrorEndpoint: "<HB_PRIVATE_REGISTRY_MIRROR_ENDPOINT>"
-
-  ##########################################
-  # STANDALONE CONFIG - INFRASTRUCTURE SETUP
-  ##########################################
-  awsCredentials:
-    awsAccessKey: "<required>"
-    awsSecretKey: "<required>"
-
-  awsConfig:
-    ami: "<HB_AWS_AMI>"
-    awsKeyName: "<HB_SSH_PRIVATE_KEY_NAME>"
-    awsInstanceType: "<HB_AWS_INSTANCE_TYPE>"
-    awsVolumeType: "<HB_AWS_VOLUME_TYPE>"
-    region: "<HB_AWS_REGION>"
-    awsSecurityGroups: ["<HB_AWS_SECURITY_GROUPS>"]
-    awsSecurityGroupNames: ["<HB_AWS_SECURITY_GROUP_NAMES>"]
-    awsSubnetID: "<HB_AWS_SUBNET_ID>"
-    awsVpcID: "<HB_AWS_VPC_ID>"
-    awsZoneLetter: "<HB_AWS_ZONE_LETTER>"
-    awsRootSize: "<HB_AWS_ROOT_SIZE>"
-    awsRoute53Zone: "<HB_AWS_ROUTE_53_ZONE>"
-    awsUser: "<HB_AWS_USER>"
-    ipAddressType: "<HB_IP_ADDRESS_TYPE>"
-    loadBalancerType: "<HB_LOAD_BALANCER_TYPE>"
-    targetType: "instance"
-    sshConnectionType: "ssh"
-    timeout: "5m"
-
-  ###################################
-  # STANDALONE CONFIG - RANCHER SETUP
-  ###################################
-  standalone:
-    bootstrapPassword: "<HB_RANCHER_ADMIN_PASSWORD>"
-    certManagerVersion: "<HB_CERT_MANAGER_VERSION>"
-    certType: "self-signed"
-    chartVersion: "<required>"
-    osUser: "<HB_OS_USER>"
-    osGroup: "<HB_OS_GROUP>"
-    registryPassword: "<HB_PRIVATE_REGISTRY_PASSWORD>"
-    registryUsername: "<HB_PRIVATE_REGISTRY_USERNAME>"
-    rancherChartRepository: "<required>"
-    rancherHostname: "<required>"
-    rancherImage: "rancher/rancher"
-    rancherTagVersion: "<required>"
-    repo: "<required>"
-    rke2Version: "<required>"
-
-terratest:
-  pathToRepo: "<HB_PATH_TO_REPO>"
-  standaloneLogging: false
-  tfLogging: false
