+++ tests/infrastructure/ranchers/internal/constants.go
@@ -1,0 +1,7 @@
+package internal
+
+const (
+	ConfigEnvironmentKey = "CATTLE_TEST_CONFIG"
+	Head                 = "head"
+	Local                = "local"
+)
