+++ tests/infrastructure/ranchers/setup/proxy/rancher.go
@@ -1,0 +1,101 @@
+package proxy
+
+import (
+	"os"
+	"testing"
+
+	"github.com/gruntwork-io/terratest/modules/terraform"
+	"github.com/rancher/shepherd/clients/rancher"
+	"github.com/rancher/shepherd/extensions/defaults/namespaces"
+	"github.com/rancher/shepherd/pkg/config/operations"
+	"github.com/rancher/shepherd/pkg/session"
+	"github.com/rancher/tests/actions/workloads/deployment"
+	"github.com/rancher/tests/actions/workloads/pods"
+	infraConfig "github.com/rancher/tests/validation/recurring/infrastructure/config"
+	"github.com/rancher/tfp-automation/config"
+	"github.com/rancher/tfp-automation/defaults/keypath"
+	"github.com/rancher/tfp-automation/defaults/providers"
+	"github.com/rancher/tfp-automation/defaults/stevetypes"
+	"github.com/rancher/tfp-automation/framework"
+	proxy "github.com/rancher/tfp-automation/framework/set/resources/proxy"
+	"github.com/rancher/tfp-automation/framework/set/resources/rancher2"
+	"github.com/rancher/tfp-automation/tests/extensions/provisioning"
+	rancherinternal "github.com/rancher/tfp-automation/tests/infrastructure/ranchers/internal"
+	ranchersetup "github.com/rancher/tfp-automation/tests/infrastructure/ranchers/setup"
+	"github.com/sirupsen/logrus"
+	"github.com/stretchr/testify/require"
+)
+
+// CreateProxyRancher is a function that creates a proxy Rancher setup, either via CLI or web application
+func CreateProxyRancher(t *testing.T, provider string, cattleConfig map[string]any) error {
+	os.Getenv("CLOUD_PROVIDER_VERSION")
+
+	rancherConfig, terraformConfig, terratestConfig, _ := config.LoadTFPConfigs(cattleConfig)
+
+	if provider != "" {
+		terraformConfig.Provider = provider
+	}
+
+	_, keyPath := rancher2.SetKeyPath(keypath.ProxyKeyPath, terratestConfig.PathToRepo, terraformConfig.Provider)
+	terraformOptions := framework.Setup(t, terraformConfig, terratestConfig, keyPath)
+
+	_, _, err := proxy.CreateMainTF(t, terraformOptions, keyPath, rancherConfig, terraformConfig, terratestConfig)
+	if err != nil {
+		return err
+	}
+
+	testSession := session.NewSession()
+
+	_, err = ranchersetup.PostRancherSetup(t, terraformOptions, rancherConfig, testSession, terraformConfig.Standalone.RancherHostname, keyPath, false)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// SetupProxyRancher sets up a proxy Rancher server and returns the client, configuration, and Terraform options.
+func SetupProxyRancher(t *testing.T, session *session.Session, moduleKeyPath string, cattleConfig map[string]any) (*rancher.Client, string, string,
+	*terraform.Options, *terraform.Options, map[string]any) {
+	var err error
+
+	rancherConfig, terraformConfig, terratestConfig, standaloneConfig := config.LoadTFPConfigs(cattleConfig)
+
+	_, keyPath := rancher2.SetKeyPath(moduleKeyPath, terratestConfig.PathToRepo, terraformConfig.Provider)
+	standaloneTerraformOptions := framework.Setup(t, terraformConfig, terratestConfig, keyPath)
+
+	proxyBastion, proxyPrivateIP, err := proxy.CreateMainTF(t, standaloneTerraformOptions, keyPath, rancherConfig, terraformConfig, terratestConfig)
+	require.NoError(t, err)
+
+	// For providers that do not have built-in DNS records, this will update the Rancher server URL.
+	if terraformConfig.Provider != providers.AWS {
+		_, err = operations.ReplaceValue([]string{"rancher", "host"}, terraformConfig.Standalone.RancherHostname, cattleConfig)
+		require.NoError(t, err)
+
+		rancherConfig, terraformConfig, terratestConfig, _ = config.LoadTFPConfigs(cattleConfig)
+		infraConfig.WriteConfigToFile(os.Getenv(rancherinternal.ConfigEnvironmentKey), cattleConfig)
+	}
+
+	client, err := ranchersetup.PostRancherSetup(t, standaloneTerraformOptions, rancherConfig, session, terraformConfig.Standalone.RancherHostname, keyPath, false)
+	require.NoError(t, err)
+
+	if standaloneConfig.RancherTagVersion != rancherinternal.Head {
+		provisioning.VerifyRancherVersion(t, rancherConfig.Host, standaloneConfig.RancherTagVersion, keyPath, standaloneTerraformOptions)
+	}
+
+	_, keyPath = rancher2.SetKeyPath(keypath.RancherKeyPath, terratestConfig.PathToRepo, "")
+	terraformOptions := framework.Setup(t, terraformConfig, terratestConfig, keyPath)
+
+	cluster, err := client.Steve.SteveType(stevetypes.Provisioning).ByID(namespaces.FleetLocal + "/local")
+	require.NoError(t, err)
+
+	logrus.Infof("Verifying cluster deployments (%s)", cluster.Name)
+	err = deployment.VerifyClusterDeployments(client, cluster)
+	require.NoError(t, err)
+
+	logrus.Infof("Verifying cluster pods (%s)", cluster.Name)
+	err = pods.VerifyClusterPods(client, cluster)
+	require.NoError(t, err)
+
+	return client, proxyBastion, proxyPrivateIP, standaloneTerraformOptions, terraformOptions, cattleConfig
+}
