+++ tests/infrastructure/ranchers/setup/token.go
@@ -1,4 +1,4 @@
-package ranchers
+package setup
 
 import (
 	"context"
