--- tests/infrastructure/ranchers/upgrading_airgap_rancher.go
@@ -1,66 +1,0 @@
-package ranchers
-
-import (
-	"os"
-	"testing"
-
-	shepherdConfig "github.com/rancher/shepherd/pkg/config"
-	"github.com/rancher/shepherd/pkg/session"
-	"github.com/rancher/tfp-automation/config"
-	"github.com/rancher/tfp-automation/defaults/keypath"
-	"github.com/rancher/tfp-automation/framework"
-	featureDefaults "github.com/rancher/tfp-automation/framework/set/defaults/features"
-	resources "github.com/rancher/tfp-automation/framework/set/resources/airgap"
-	"github.com/rancher/tfp-automation/framework/set/resources/rancher2"
-	"github.com/rancher/tfp-automation/framework/set/resources/upgrade"
-	"github.com/rancher/tfp-automation/tests/extensions/ssh"
-	"github.com/stretchr/testify/require"
-)
-
-// UpgradingAirgapRancher is a function that creates and upgrades an airgap Rancher setup, either via CLI or web application
-func UpgradingAirgapRancher(t *testing.T, provider string) error {
-	os.Getenv("CLOUD_PROVIDER_VERSION")
-
-	configPath := os.Getenv("CATTLE_TEST_CONFIG")
-	cattleConfig := shepherdConfig.LoadConfigFromFile(configPath)
-	rancherConfig, terraformConfig, terratestConfig, standaloneConfig := config.LoadTFPConfigs(cattleConfig)
-
-	if provider != "" {
-		terraformConfig.Provider = provider
-	}
-
-	_, keyPath := rancher2.SetKeyPath(keypath.AirgapKeyPath, terratestConfig.PathToRepo, terraformConfig.Provider)
-	terraformOptions := framework.Setup(t, terraformConfig, terratestConfig, keyPath)
-
-	registry, bastion, err := resources.CreateMainTF(t, terraformOptions, keyPath, rancherConfig, terraformConfig, terratestConfig)
-	require.NoError(t, err)
-
-	sshKey, err := os.ReadFile(terraformConfig.PrivateKeyPath)
-	require.NoError(t, err)
-
-	_, err = ssh.StartBastionSSHTunnel(bastion, terraformConfig.Standalone.OSUser, sshKey, "8443", standaloneConfig.RancherHostname, "443")
-	require.NoError(t, err)
-
-	testSession := session.NewSession()
-
-	client, err := PostRancherSetup(t, terraformOptions, rancherConfig, testSession, terraformConfig.Standalone.RancherHostname, keyPath, false)
-	require.NoError(t, err)
-
-	if standaloneConfig.FeatureFlags != nil && standaloneConfig.FeatureFlags.Turtles != "" {
-		toggleFeatureFlag(client, featureDefaults.Turtles, standaloneConfig.FeatureFlags.Turtles)
-	}
-
-	terraformConfig.Standalone.UpgradeAirgapRancher = true
-
-	_, upgradeKeyPath := rancher2.SetKeyPath(keypath.UpgradeKeyPath, terratestConfig.PathToRepo, terraformConfig.Provider)
-	upgradeTerraformOptions := framework.Setup(t, terraformConfig, terratestConfig, upgradeKeyPath)
-
-	err = upgrade.CreateMainTF(t, upgradeTerraformOptions, upgradeKeyPath, rancherConfig, terraformConfig, terratestConfig, "", "", bastion, registry)
-	require.NoError(t, err)
-
-	standaloneTerraformOptions := framework.Setup(t, terraformConfig, terratestConfig, keypath.AirgapKeyPath)
-	client, err = PostRancherSetup(t, standaloneTerraformOptions, rancherConfig, testSession, terraformConfig.Standalone.RancherHostname, keyPath, true)
-	require.NoError(t, err)
-
-	return nil
-}
