+++ tests/rancher2/recurring/createRancherServer.go
@@ -12,7 +12,8 @@ import (
 	"github.com/rancher/shepherd/pkg/session"
 	infraConfig "github.com/rancher/tests/validation/recurring/infrastructure/config"
 	"github.com/rancher/tfp-automation/defaults/keypath"
-	"github.com/rancher/tfp-automation/tests/infrastructure/ranchers"
+
+	setupstandard "github.com/rancher/tfp-automation/tests/infrastructure/ranchers/setup/standard"
 	"github.com/sirupsen/logrus"
 )
 
@@ -24,7 +25,7 @@ func main() {
 
 	cattleConfig := shepherdConfig.LoadConfigFromFile(os.Getenv(shepherdConfig.ConfigEnvironmentKey))
 
-	client, _, _, _, _, err = setupRancher(t)
+	client, _, _, _, _, err = setupRancher(t, cattleConfig)
 	if err != nil {
 		logrus.Fatalf("Failed to setup Rancher: %v", err)
 	}
@@ -37,9 +38,9 @@ func main() {
 	infraConfig.WriteConfigToFile(os.Getenv(config.ConfigEnvironmentKey), cattleConfig)
 }
 
-func setupRancher(t *testing.T) (*rancher.Client, string, *terraform.Options, *terraform.Options, map[string]any, error) {
+func setupRancher(t *testing.T, cattleConfig map[string]any) (*rancher.Client, string, *terraform.Options, *terraform.Options, map[string]any, error) {
 	testSession := session.NewSession()
-	client, serverNodeOne, standaloneTerraformOptions, terraformOptions, cattleConfig := ranchers.SetupRancher(t, testSession, keypath.SanityKeyPath)
+	client, serverNodeOne, standaloneTerraformOptions, terraformOptions, returnCattleConfig := setupstandard.SetupRancher(t, testSession, keypath.SanityKeyPath, cattleConfig)
 
-	return client, serverNodeOne, standaloneTerraformOptions, terraformOptions, cattleConfig, nil
+	return client, serverNodeOne, standaloneTerraformOptions, terraformOptions, returnCattleConfig, nil
 }
