+++ tests/sanity/sanity_upgrade_ipv6_rancher_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/gruntwork-io/terratest/modules/terraform"
 	"github.com/rancher/shepherd/clients/rancher"
+	shepherdConfig "github.com/rancher/shepherd/pkg/config"
 	"github.com/rancher/shepherd/pkg/session"
 	clusterActions "github.com/rancher/tests/actions/clusters"
 	provisioningActions "github.com/rancher/tests/actions/provisioning"
@@ -20,7 +21,10 @@ import (
 	"github.com/rancher/tfp-automation/pipeline/qase/results"
 	nested "github.com/rancher/tfp-automation/tests/extensions/nestedModules"
 	"github.com/rancher/tfp-automation/tests/extensions/provisioning"
-	"github.com/rancher/tfp-automation/tests/infrastructure/ranchers"
+
+	setupipv6 "github.com/rancher/tfp-automation/tests/infrastructure/ranchers/setup/ipv6"
+	upgradebase "github.com/rancher/tfp-automation/tests/infrastructure/ranchers/upgrade"
+	upgradeipv6 "github.com/rancher/tfp-automation/tests/infrastructure/ranchers/upgrade/ipv6"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
@@ -52,28 +56,29 @@ func (s *TfpSanityIPv6UpgradeRancherTestSuite) TearDownSuite() {
 func (s *TfpSanityIPv6UpgradeRancherTestSuite) SetupSuite() {
 	testSession := session.NewSession()
 	s.session = testSession
+	s.cattleConfig = shepherdConfig.LoadConfigFromFile(os.Getenv(shepherdConfig.ConfigEnvironmentKey))
 
-	s.client, s.serverNodeOne, s.standaloneTerraformOptions, s.terraformOptions, s.cattleConfig = ranchers.SetupIPv6Rancher(s.T(), s.session, keypath.IPv6KeyPath)
+	s.client, s.serverNodeOne, s.standaloneTerraformOptions, s.terraformOptions, s.cattleConfig = setupipv6.SetupIPv6Rancher(s.T(), s.session, keypath.IPv6KeyPath, s.cattleConfig)
 	s.rancherConfig, s.terraformConfig, s.terratestConfig, s.standaloneConfig = config.LoadTFPConfigs(s.cattleConfig)
 }
 
 func (s *TfpSanityIPv6UpgradeRancherTestSuite) TestTfpUpgradeIPv6Rancher() {
-	standardUserClient, standardToken, testUser, testPassword := ranchers.SetupResources(s.T(), s.client, s.rancherConfig, s.terratestConfig, s.terraformOptions)
+	standardUserClient, standardToken, testUser, testPassword := upgradebase.SetupResources(s.T(), s.client, s.rancherConfig, s.terratestConfig, s.terraformOptions)
 
 	s.rancherConfig, s.terraformConfig, s.terratestConfig, _ = config.LoadTFPConfigs(s.cattleConfig)
 	nestedRancherModuleDir := s.provisionAndVerifyCluster("Sanity_IPv6_Pre_Rancher_Upgrade", standardUserClient, standardToken, testUser, testPassword)
 
-	s.client, s.cattleConfig, s.terraformOptions, s.upgradeTerraformOptions = ranchers.UpgradeIPv6Rancher(s.T(), s.client, s.serverNodeOne, s.session, s.cattleConfig)
-
-	ranchers.CleanupDownstreamClusters(s.T(), s.client, s.terraformConfig)
+	s.client, s.cattleConfig, s.terraformOptions, s.upgradeTerraformOptions = upgradeipv6.UpgradeIPv6Rancher(s.T(), s.client, s.serverNodeOne, s.session, s.cattleConfig)
+	upgradebase.
+		CleanupDownstreamClusters(s.T(), s.client, s.terraformConfig)
 	os.RemoveAll(nestedRancherModuleDir)
 
-	standardUserClient, standardToken, testUser, testPassword = ranchers.SetupResources(s.T(), s.client, s.rancherConfig, s.terratestConfig, s.terraformOptions)
+	standardUserClient, standardToken, testUser, testPassword = upgradebase.SetupResources(s.T(), s.client, s.rancherConfig, s.terratestConfig, s.terraformOptions)
 
 	s.rancherConfig, s.terraformConfig, s.terratestConfig, _ = config.LoadTFPConfigs(s.cattleConfig)
 	nestedRancherModuleDir = s.provisionAndVerifyCluster("Sanity_IPv6_Post_Rancher_Upgrade", standardUserClient, standardToken, testUser, testPassword)
-
-	ranchers.CleanupDownstreamClusters(s.T(), s.client, s.terraformConfig)
+	upgradebase.
+		CleanupDownstreamClusters(s.T(), s.client, s.terraformConfig)
 	os.RemoveAll(nestedRancherModuleDir)
 
 	if s.terratestConfig.LocalQaseReporting {
